Fixed pedantic C errors.

+++ test/test_network.cpp
@@ -43,12 +43,12 @@ void testNetworkMatrix(
     ASSERT_CMR_CALL( CMRchrmatPrintDense(cmr, result, stdout, '0', true) );
 
     printf("Basis:");
-    for (int r = 0; r < matrix->numRows; ++r)
+    for (size_t r = 0; r < matrix->numRows; ++r)
       printf(" %d", basis[r]);
     printf("\n");
 
     printf("Cobasis:");
-    for (int c = 0; c < matrix->numColumns; ++c)
+    for (size_t c = 0; c < matrix->numColumns; ++c)
       printf(" %d", cobasis[c]);
     printf("\n");
   }
